refatoração: Movido código de DailyLiturgyViewModel para LiturgyCardView
- Removido o código de LiturgyCard do DailyLiturgyViewModel
- Criado novo arquivo LiturgyCardView
